+++ docs/changelog.md
@@ -1,3 +1,20 @@
+<<<<<<< HEAD
+<<<<<<< HEAD
+---
+title: "Changelog"
+module: "Media"
+type: concept
+tags: [changelog]
+created: 2026-07-14
+updated: 2026-07-14
+qmd: "changelog"
+related:
+  - "./webm.md"
+---
+=======
+>>>>>>> f6dc2a0 (.)
+=======
+>>>>>>> 11477b67d (.)
 # Changelog
 
 All notable changes to this project will be documented in this file.